requirements.txt
streamlit
pandas
sqlalchemy
psycopg2
streamlit-extras
streamlit-authenticator
plotly
entrypoints
streamlit-option-menu
pygwalker
streamlit-aggrid
mysql-connector-python
mysql-connector
python-dotenv
deta